OneDrive/Desktop/MTN/External Projects/UZ - Survey and Geomatics/Agriculture - ML model/requirements.txt
flask
scikit-learn
pandas
numpy
matplotlib
seaborn
folium
rasterio
geopandas
shapely
joblib
scipy
requests
earthengine-api
google-auth
pyproj
cryptography
python-dotenv
gunicorn
